Refactor call tree rendering responsibilities
