+++ src/ProfilerCore/Rendering/ProfilerCallTreeFormatter.cs
@@ -1,24 +1,23 @@
 using System;
 using System.Globalization;
-using System.Text.RegularExpressions;
 using Spectre.Console;
 using static Asynkron.Profiler.CallTreeHelpers;
 
 namespace Asynkron.Profiler;
 
-internal sealed partial class ProfilerCallTreeFormatter
+internal sealed class ProfilerCallTreeFormatter
 {
-    private const double HotnessColorFloor = 0.001d;
-    private const double HotnessColorMid = 0.0025d;
-    private const double HotnessColorMax = 0.4d;
     private const string HotspotMarker = "\U0001F525";
 
     private readonly Theme _theme;
-    private readonly Regex _jitNumberRegex = JitNumberRegex();
+    private readonly ProfilerHotnessColorScale _hotnessColorScale;
+    private readonly ProfilerJitNumberHighlighter _jitNumberHighlighter;
 
     public ProfilerCallTreeFormatter(Theme theme)
     {
         _theme = theme;
+        _hotnessColorScale = new ProfilerHotnessColorScale(theme);
+        _jitNumberHighlighter = new ProfilerJitNumberHighlighter(theme);
     }
 
     public string FormatAllocationCallTreeLine(
@@ -77,7 +76,7 @@ internal sealed partial class ProfilerCallTreeFormatter
 
         var truncatedName = Truncate(displayName, maxNameLength);
         var hotness = ComputeHotness(node, totalTime, totalSamples);
-        var nameColor = useHeatColor ? GetHotnessColor(hotness) : null;
+        var nameColor = useHeatColor ? _hotnessColorScale.GetColor(hotness) : null;
         var nameText = FormatCallTreeName(truncatedName, matchName, ShouldStopAtLeaf(matchName), nameColor);
 
         var visibleLength = statsLength + truncatedName.Length;
@@ -110,7 +109,7 @@ internal sealed partial class ProfilerCallTreeFormatter
         var pctText = pct.ToString("F1", CultureInfo.InvariantCulture);
         var countText = node.Calls.ToString("N0", CultureInfo.InvariantCulture) + countSuffix;
         var hotness = ComputeHotness(node, totalTime, totalSamples);
-        var nameColor = useHeatColor ? GetHotnessColor(hotness) : null;
+        var nameColor = useHeatColor ? _hotnessColorScale.GetColor(hotness) : null;
 
         var maxNameLength = timeline != null
             ? Math.Max(15, timeline.TextWidth - depth * 4 - $"{timeText} {timeUnitLabel} {pctText}% {countText} ".Length)
@@ -128,7 +127,7 @@ internal sealed partial class ProfilerCallTreeFormatter
             return baseLine;
         }
 
-        var bar = RenderTimelineBar(node, timeline);
+        var bar = CallTreeTimelineBarRenderer.Render(node, timeline);
         var visibleLength = EstimateVisibleLength(baseLine);
         var padding = timeline.GetPaddingForDepth(depth, visibleLength);
         return $"{baseLine}{new string(' ', padding)} [dim]│[/] {bar}";
@@ -171,68 +170,8 @@ internal sealed partial class ProfilerCallTreeFormatter
 
     public string HighlightJitNumbers(string text)
     {
-        return _jitNumberRegex.Replace(text, match => WrapAnsi(match.Value, _theme.AccentColor));
+        return _jitNumberHighlighter.Highlight(text);
     }
-
-    public static string RenderTimelineBar(CallTreeNode node, TimelineContext context)
-    {
-        if (!node.HasTiming || context.RootDuration <= 0)
-        {
-            return new string(' ', context.BarWidth);
-        }
-
-        var buffer = new char[context.BarWidth];
-        Array.Fill(buffer, ' ');
-
-        var startOffset = node.MinStart - context.RootStart;
-        var startRatio = Math.Clamp(startOffset / context.RootDuration, 0, 1);
-        var durationRatio = Math.Clamp((node.MaxEnd - node.MinStart) / context.RootDuration, 0, 1);
-
-        var startPosition = startRatio * context.BarWidth;
-        var endPosition = (startRatio + durationRatio) * context.BarWidth;
-
-        var leftIndex = (int)Math.Floor(startPosition);
-        var rightIndex = (int)Math.Ceiling(endPosition) - 1;
-        leftIndex = Math.Clamp(leftIndex, 0, context.BarWidth - 1);
-        rightIndex = Math.Clamp(rightIndex, 0, context.BarWidth - 1);
-
-        if (rightIndex < leftIndex)
-        {
-            rightIndex = leftIndex;
-        }
-
-        if (leftIndex == rightIndex)
-        {
-            var singleFraction = Math.Clamp(endPosition - startPosition, 0, 1);
-            buffer[leftIndex] = singleFraction switch
-            {
-                >= 0.875 => '█',
-                >= 0.625 => '▊',
-                >= 0.375 => '▌',
-                >= 0.125 => '▎',
-                _ => ' '
-            };
-        }
-        else
-        {
-            var leftFraction = 1 - (startPosition - leftIndex);
-            buffer[leftIndex] = SelectLeftBlock(leftFraction);
-
-            for (var index = leftIndex + 1; index < rightIndex; index++)
-            {
-                buffer[index] = '█';
-            }
-
-            var rightFraction = endPosition - Math.Floor(endPosition);
-            buffer[rightIndex] = rightFraction <= 0 ? '█' : SelectRightBlock(rightFraction);
-        }
-
-        var heat = Math.Clamp((node.MaxEnd - node.MinStart) / context.RootDuration, 0, 1);
-        var color = GetHeatColor(heat);
-        return $"[{color}]{Markup.Escape(new string(buffer))}[/]";
-    }
-
-    public static bool IsFireEmojiCandidate(double hotness, double hotThreshold) => hotness >= hotThreshold;
 
     private static double GetDisplayedTime(CallTreeNode node, bool useSelfTime, bool isRoot)
     {
@@ -268,16 +207,6 @@ internal sealed partial class ProfilerCallTreeFormatter
         return value.Length > maxLength ? value[..(maxLength - 3)] + "..." : value;
     }
 
-    private static string WrapAnsi(string text, string? color)
-    {
-        if (string.IsNullOrWhiteSpace(color))
-        {
-            return text;
-        }
-
-        return $"[{color}]{Markup.Escape(text)}[/]";
-    }
-
     private static int EstimateVisibleLength(string markup)
     {
         var result = markup;
@@ -300,90 +229,4 @@ internal sealed partial class ProfilerCallTreeFormatter
 
         return result.Length;
     }
-
-    private string? GetHotnessColor(double hotness)
-    {
-        if (!ConsoleThemeHelpers.TryParseHexColor(_theme.TextColor, out var cool) ||
-            !ConsoleThemeHelpers.TryParseHexColor(_theme.HotColor, out var hot))
-        {
-            return null;
-        }
-
-        double normalizedHotness;
-        if (hotness <= HotnessColorFloor)
-        {
-            normalizedHotness = 0d;
-        }
-        else if (hotness <= HotnessColorMid)
-        {
-            var span = HotnessColorMid - HotnessColorFloor;
-            normalizedHotness = span > 0d
-                ? (hotness - HotnessColorFloor) / span * 0.5d
-                : 0d;
-        }
-        else if (hotness >= HotnessColorMax)
-        {
-            normalizedHotness = 1d;
-        }
-        else
-        {
-            var span = HotnessColorMax - HotnessColorMid;
-            normalizedHotness = span > 0d
-                ? 0.5d + (hotness - HotnessColorMid) / span * 0.5d
-                : 1d;
-        }
-
-        normalizedHotness = Math.Clamp(normalizedHotness, 0d, 1d);
-        return InterpolateColor(cool, hot, normalizedHotness);
-    }
-
-    private static string InterpolateColor((byte R, byte G, byte B) start, (byte R, byte G, byte B) end, double t)
-    {
-        t = Math.Clamp(t, 0d, 1d);
-        var r = (byte)Math.Round(start.R + (end.R - start.R) * t);
-        var g = (byte)Math.Round(start.G + (end.G - start.G) * t);
-        var b = (byte)Math.Round(start.B + (end.B - start.B) * t);
-        return $"#{r:X2}{g:X2}{b:X2}";
-    }
-
-    private static string GetHeatColor(double percentage)
-    {
-        return percentage switch
-        {
-            >= 0.75 => "red",
-            >= 0.50 => "orange1",
-            >= 0.25 => "yellow1",
-            _ => "grey"
-        };
-    }
-
-    private static char SelectLeftBlock(double fraction)
-    {
-        return fraction switch
-        {
-            >= 1.0 => '█',
-            >= 0.875 => '▉',
-            >= 0.75 => '▊',
-            >= 0.625 => '▋',
-            >= 0.5 => '▌',
-            >= 0.375 => '▍',
-            >= 0.25 => '▎',
-            >= 0.125 => '▏',
-            _ => ' '
-        };
-    }
-
-    private static char SelectRightBlock(double fraction)
-    {
-        return fraction switch
-        {
-            >= 1.0 => '█',
-            >= 0.5 => '▐',
-            >= 0.125 => '▕',
-            _ => ' '
-        };
-    }
-
-    [GeneratedRegex(@"(?<![A-Za-z0-9_])(#?0x[0-9A-Fa-f]+|#?\d+)(?![A-Za-z0-9_])", RegexOptions.Compiled)]
-    private static partial Regex JitNumberRegex();
 }
